Type truncation warning fixes in a few networking classes [CODEREVIEW] ryan.gerleve #jira UE-116218
#ROBOMERGE-AUTHOR: michael.noland
#ROBOMERGE-SOURCE: CL 17578561 in //UE5/Main/... via CL 17578567
#ROBOMERGE-BOT: STARSHIP (Release-Engine-Test -> Release-Engine-Staging) (v871-17566257)

[CL 17578577 by michael noland in ue5-release-engine-staging branch]

+++ Engine/Source/Runtime/CoreUObject/Public/UObject/CoreNet.h
@@ -282,7 +282,7 @@ public:
 	}
 
 	FLifetimeProperty(int32 InRepIndex)
-		: RepIndex(InRepIndex)
+		: RepIndex((uint16)InRepIndex)
 		, Condition(COND_None)
 		, RepNotifyCondition(REPNOTIFY_OnChanged)
 		, bIsPushBased(false)
@@ -291,7 +291,7 @@ public:
 	}
 
 	FLifetimeProperty(int32 InRepIndex, ELifetimeCondition InCondition, ELifetimeRepNotifyCondition InRepNotifyCondition=REPNOTIFY_OnChanged, bool bInIsPushBased=false)
-		: RepIndex(InRepIndex)
+		: RepIndex((uint16)InRepIndex)
 		, Condition(InCondition)
 		, RepNotifyCondition(InRepNotifyCondition)
 		, bIsPushBased(bInIsPushBased)
